Data import utilities

In order to speed up data access operations, sepsr does not directly use the PhysioNet provided CSV files, but converts all data to fst::fst() format, which allows for random row and column access. Large tables are split into chunks in order to keep memory requirements under control.

import_mimic(demo = FALSE, dir = mimic_data_dir(demo),
  config = mimic_config(demo), cleanup = TRUE)

import_eicu(demo = FALSE, dir = eicu_data_dir(demo),
  config = eicu_config(demo), cleanup = TRUE)

import_datasource(dir, config, cleanup = TRUE)

Arguments

demo
Logical flag for switching between default arguments for demo or full data sources.

dir
The directory where the data was downloaded to (see download_mimic()).

config
Import configuration specified as nested list.

cleanup
Logical flag, if TRUE the original CSV data is deleted after successful import.

Details

A nested list supplied as config argument controls how the CSV data is parsed and ingested. A the top level, elements name (string), base_url (string), version (string) and tables (list) are expected. For every file to be read in, an entry in tables with the name of the corresponding file is assumed, requiring a list named col_spec and optionally a list named partitioning.

Every col_spec node again is expected to be a named list with names corresponding to column names in the CSV and list valued entries. Finally, a list node defining a single column is a named list with a required, string-valued type entry which is interpreted as readr column specification function (see readr::cols()). Further list entries will be passed as arguments to the column specification function. A column defined as

list(type = "col_datetime", format = "%Y-%m-%d %H:%M:%S")

will therefore result in a readr column specification like

col_datetime(format = "%Y-%m-%d %H:%M:%S")

Entries controlling the partitioning of tables, specified as partitioning list nodes are expected to be named lists of length 1, where the name corresponds to the column used for partitioning and partitions are defined as integer vector which is passed to base::findInterval(). This implies that partitions can only be created on integer valued columns.

Putting all this together, the expected structure of config is as

list(
  name = "some_name",
  base_url = "some_url",
  version = "major.minor",
  tables = list(
    file_name_table_1 = list(
      col_spec = list(
        colname_1 = list(type = "col_*"),
        colname_2 = list(type = "col_*"),
        ...
      )
    ),
    file_name_table_2 = list(
      col_spec = list(...),
      partitioning = list(colname_i = c(0L, ))
    ),
    ...
  )
)

Configurations for the MIMIC-III and eICU databases are available in extdata/config as JSON files and can be read in using get_config().

Examples

if (FALSE) {

dir <- tempdir()
list.files(dir)

download_mimic(demo = TRUE, dest = dir)
list.files(dir)

import_mimic(demo = TRUE, dir = dir)
list.files(dir)

unlink(dir, recursive = TRUE)

}